Introduce basic workload scheduler module
